refactor: Modals -> move into shared UI directory
